fix: add `includeSummaryInBody` flag from SLA
In addressing `SCHOOL-108 Section pages have blank titles` it was discovered that the
blank title issue had already been addressed but SLA had previous overides so were not
seeing the fix.  These overrides introduced the `includeSummaryInBody` flag for blog posts.
It was decided that the SLA behavior should be preserved and applied to all schools.